requirements.txt
streamlit
pandas
python-dotenv
langchain-groq
tavily-python
numpy<2